app/requirements.txt
boto3==1.26.115
botocore==1.29.115
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Faker==18.4.0
Flask==2.2.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
mdgen==0.1.10
Pillow==9.5.0
psycopg2-binary==2.9.5
pyshorteners==1.0.1
requests==2.28.2
s3transfer==0.6.0
urllib3==1.26.15
Werkzeug==2.2.3
gunicorn==20.1.0
